ai/Injection Everywhere/src/requirements.txt
flask
requests
google-generativeai
python-dotenv
